+++ packages/desktop/CHANGELOG.md
@@ -3,6 +3,94 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.7-alpha.2](https://github.com/TryQuiet/monorepo/compare/quiet@0.9.7-alpha.1...quiet@0.9.7-alpha.2) (2022-07-24)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+## [0.9.7-alpha.1](https://github.com/TryQuiet/monorepo/compare/quiet@0.9.7-alpha.0...quiet@0.9.7-alpha.1) (2022-07-23)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+## [0.9.7-alpha.0](https://github.com/ZbayApp/monorepo/compare/quiet@0.8.6...quiet@0.9.7-alpha.0) (2022-07-15)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+## [0.9.6-alpha.0](https://github.com/ZbayApp/monorepo/compare/quiet@0.9.5-alpha.1...quiet@0.9.6-alpha.0) (2022-07-15)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+## [0.9.5-alpha.1](https://github.com/ZbayApp/monorepo/compare/quiet@0.9.5-alpha.0...quiet@0.9.5-alpha.1) (2022-07-15)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+## [0.9.5-alpha.0](https://github.com/ZbayApp/monorepo/compare/quiet@0.9.4-alpha.0...quiet@0.9.5-alpha.0) (2022-07-07)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+## [0.9.4-alpha.0](https://github.com/ZbayApp/monorepo/compare/quiet@0.9.3-alpha.0...quiet@0.9.4-alpha.0) (2022-07-06)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+## [0.9.3-alpha.0](https://github.com/ZbayApp/monorepo/compare/quiet@0.9.2-alpha.0...quiet@0.9.3-alpha.0) (2022-07-06)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+## [0.9.2-alpha.0](https://github.com/ZbayApp/monorepo/compare/quiet@0.9.1-alpha.0...quiet@0.9.2-alpha.0) (2022-07-05)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+## [0.9.1-alpha.0](https://github.com/ZbayApp/monorepo/compare/quiet@0.9.0-alpha.0...quiet@0.9.1-alpha.0) (2022-07-05)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [0.9.0-alpha.0](https://github.com/ZbayApp/monorepo/compare/quiet@0.8.6...quiet@0.9.0-alpha.0) (2022-07-05)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
 ## [0.8.6](https://github.com/TryQuiet/monorepo/compare/quiet@0.8.5...quiet@0.8.6) (2022-06-23)
 
 **Note:** Version bump only for package quiet